Move more headers
