+++ src/ngraph_var.h
